# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/request-otp": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOtpResponseDto"
                }
              }
            }
          }
        },
        "summary": "전화번호로 인증번호(OTP) 발급 — 목 OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResponseDto"
                }
              }
            }
          }
        },
        "summary": "인증번호 검증 — 기존 유저면 토큰, 신규면 가입 진행 신호",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "인증 완료 후 생일·닉네임으로 회원가입",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "현재 인증된 유저 조회",
        "tags": [
          "auth"
        ]
      }
    },
    "/choices": {
      "get": {
        "description": "questionId 없으면 새 질문(초기/스킵), 있으면 같은 질문 + 새 후보(다시 섞기).",
        "operationId": "ChoiceController_getFeed",
        "parameters": [
          {
            "name": "questionId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChoiceFeed"
                }
              }
            }
          }
        },
        "summary": "초이스 피드 (질문 + 랜덤 후보 4명)",
        "tags": [
          "choices"
        ]
      }
    },
    "/choices/select": {
      "post": {
        "description": "로그인한 \"나\"를 selector로 기록 → 받은픽 데이터의 출처.",
        "operationId": "ChoiceController_select",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectChoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChoiceFeed"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "카드 선택 기록 후 다음 피드 반환",
        "tags": [
          "choices"
        ]
      }
    },
    "/viewer": {
      "get": {
        "operationId": "ViewerController_getViewer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Viewer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "현재 유저(\"나\") 조회 (프리미엄 여부 포함)",
        "tags": [
          "viewer"
        ]
      }
    },
    "/viewer/premium": {
      "post": {
        "operationId": "ViewerController_subscribePremium",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Viewer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "프리미엄 구독('가입하기') — 즉시 프리미엄 ON(임시)",
        "tags": [
          "viewer"
        ]
      }
    },
    "/received-picks": {
      "get": {
        "description": "비프리미엄이면 selector 사진은 서버에서 가려진다(photoUrl=null).",
        "operationId": "ReceivedPicksController_getReceivedPicks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivedPicks"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "나를 픽한 사람 목록 + 각자의 받은픽 Top3",
        "tags": [
          "received-picks"
        ]
      }
    },
    "/conversations": {
      "get": {
        "operationId": "ConversationsController_getConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "소통 목록(메시지함)",
        "tags": [
          "conversations"
        ]
      },
      "post": {
        "operationId": "ConversationsController_createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "소통하기 — 대화 생성(이미 있으면 반환)",
        "tags": [
          "conversations"
        ]
      }
    },
    "/conversations/{id}/messages": {
      "get": {
        "operationId": "ConversationsController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메시지 히스토리 (열면 읽음 처리)",
        "tags": [
          "conversations"
        ]
      },
      "post": {
        "operationId": "ConversationsController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메시지 전송 (REST 폴백; 정식 송신은 소켓)",
        "tags": [
          "conversations"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Liveness/readiness check (app + DB)",
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "Feel Pick API",
    "description": "Pick & vote API",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "010-1234-5678",
            "description": "전화번호(하이픈 허용)"
          }
        },
        "required": [
          "phone"
        ]
      },
      "RequestOtpResponseDto": {
        "type": "object",
        "properties": {
          "devCode": {
            "type": "string",
            "nullable": true,
            "description": "개발 환경에서만 채워지는 인증번호(목 OTP). 운영에선 null.",
            "example": "123456"
          }
        },
        "required": [
          "devCode"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "010-1234-5678",
            "description": "전화번호(하이픈 허용)"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6자리 인증번호"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "VerifyOtpResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true,
            "description": "기존 유저면 발급된 토큰, 신규 유저면 null(가입 진행 필요)."
          },
          "isNewUser": {
            "type": "boolean",
            "description": "가입 이력이 없는 신규 유저인지 여부."
          }
        },
        "required": [
          "accessToken",
          "isNewUser"
        ]
      },
      "SignupDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "010-1234-5678",
            "description": "전화번호(하이픈 허용)"
          },
          "birthday": {
            "type": "string",
            "example": "2000-09-20",
            "format": "date",
            "description": "생년월일 (ISO 날짜)"
          },
          "nickname": {
            "type": "string",
            "example": "아니근데옥지얌",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": [
          "phone",
          "birthday",
          "nickname"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string",
            "description": "전화번호(정규화된 숫자만)",
            "example": "01012345678"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "생년월일"
          },
          "displayName": {
            "type": "string"
          },
          "isPremium": {
            "type": "boolean",
            "description": "프리미엄 구독 여부(임시)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "phone",
          "birthday",
          "displayName",
          "isPremium",
          "createdAt"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "example": "술 잘 먹을 것 같은 친구"
          }
        },
        "required": [
          "id",
          "text"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "하리니"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://i.pravatar.cc/600?img=5"
          },
          "distanceKm": {
            "type": "number",
            "nullable": true,
            "example": 17,
            "description": "나와의 거리(km, 목업)"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "example": "@hx_rxx_ 맞팔도 받아용!"
          },
          "interests": {
            "example": [
              "홍대",
              "스티커 사진",
              "닌텐도"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "photoUrl",
          "distanceKm",
          "bio",
          "interests"
        ]
      },
      "ChoiceFeed": {
        "type": "object",
        "properties": {
          "question": {
            "$ref": "#/components/schemas/Question"
          },
          "candidates": {
            "description": "랜덤하게 주어지는 후보 카드",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        },
        "required": [
          "question",
          "candidates"
        ]
      },
      "SelectChoiceDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "format": "uuid",
            "description": "제시된 질문 ID"
          },
          "selectedUserId": {
            "type": "string",
            "format": "uuid",
            "description": "선택한 유저 ID"
          }
        },
        "required": [
          "questionId",
          "selectedUserId"
        ]
      },
      "Viewer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "나"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://i.pravatar.cc/600?img=8"
          },
          "isPremium": {
            "type": "boolean",
            "description": "프리미엄 구독 여부(임시)"
          }
        },
        "required": [
          "id",
          "displayName",
          "photoUrl",
          "isPremium"
        ]
      },
      "Top3Item": {
        "type": "object",
        "properties": {
          "questionText": {
            "type": "string",
            "example": "술 잘 먹을 것 같은 친구"
          },
          "votes": {
            "type": "number",
            "example": 12,
            "description": "그 주제로 받은 픽 수"
          }
        },
        "required": [
          "questionText",
          "votes"
        ]
      },
      "ReceivedPick": {
        "type": "object",
        "properties": {
          "selector": {
            "description": "나를 픽한 사람 (비프리미엄이면 photoUrl=null)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Profile"
              }
            ]
          },
          "questionId": {
            "type": "string",
            "format": "uuid",
            "description": "나를 픽한 주제의 질문 ID"
          },
          "questionText": {
            "type": "string",
            "example": "1프로라도 관심이 가는 친구",
            "description": "나를 픽한 주제(대표: 가장 최근 픽)"
          },
          "pickedAt": {
            "type": "string",
            "format": "date-time"
          },
          "top3": {
            "description": "그 사람이 받은 픽 Top3",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Top3Item"
            }
          }
        },
        "required": [
          "selector",
          "questionId",
          "questionText",
          "pickedAt",
          "top3"
        ]
      },
      "ReceivedPicks": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 174,
            "description": "내가 받은 총 픽 수"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedPick"
            }
          }
        },
        "required": [
          "total",
          "items"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "example": "안녕하세요! 픽 감사해요 ㅎㅎ"
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "읽은 시각 (null이면 안읽음)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "conversationId",
          "senderId",
          "text",
          "readAt",
          "createdAt"
        ]
      },
      "ConversationSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "partner": {
            "description": "상대",
            "allOf": [
              {
                "$ref": "#/components/schemas/Profile"
              }
            ]
          },
          "questionText": {
            "type": "string",
            "nullable": true,
            "description": "소통 시작의 픽 주제(뱃지용)"
          },
          "lastMessage": {
            "nullable": true,
            "description": "마지막 메시지 (없으면 null)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Message"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "example": 2,
            "description": "안읽은 메시지 수"
          }
        },
        "required": [
          "id",
          "partner",
          "questionText",
          "lastMessage",
          "unreadCount"
        ]
      },
      "CreateConversationDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "format": "uuid",
            "description": "소통할 상대 유저 ID"
          },
          "questionId": {
            "type": "string",
            "format": "uuid",
            "description": "소통 시작의 픽 주제(뱃지용)"
          }
        },
        "required": [
          "targetUserId"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "안녕하세요!",
            "maxLength": 1000
          }
        },
        "required": [
          "text"
        ]
      }
    }
  }
}